requirements.txt
PyYAML>=5.3.1
requests>=2.24.0
Jinja2>=2.11.2
packaging>=20.4
click>=7.1.2